:recycle: Use ProcessInstanceStateHandlerFacade in ResumeProcessService
