+++ src/runtime/facades/iprocess_instance_config.ts
@@ -1,3 +1,4 @@
+import {FlowNodeInstance} from '@process-engine/flow_node_instance.contracts';
 import {IProcessModelFacade, IProcessTokenFacade, ProcessToken} from '@process-engine/process_engine_contracts';
 import {Model} from '@process-engine/process_model.contracts';
 
@@ -8,9 +9,10 @@ export interface IProcessInstanceConfig {
   correlationId: string;
   processModelId: string;
   processInstanceId: string;
-  parentProcessInstanceId: string;
+  parentProcessInstanceId?: string;
   processModelFacade: IProcessModelFacade;
   startEvent: Model.Events.StartEvent;
+  startEventInstance?: FlowNodeInstance;
   processToken: ProcessToken;
   processTokenFacade: IProcessTokenFacade;
 }
